+++ src/components/filters/option-picker/option-picker.js
@@ -52,5 +52,3 @@ OptionPicker.defaultProps = {
 };
 
 export default OptionPicker;
-
-/// /// ///
